tx_gateway_frontend: fix crash on expire_old_txs
ntp_meta->get_assignments() returns a reference to a collection and
expire_old_txs iterates it asynchronously so there is a chance of
the collection being modified out of expire_old_txs leading to the
iterator invalidation and the crash.

Fixing the probelm by making a copy before iterating.

Fixes https://github.com/redpanda-data/redpanda/issues/12503