r/heartbeats: introduce new heartbeat reply/request pair
Introduced new heartbeat types supporting lightweight heartbeats.

New versions of heartbeat request/response are optimized to make
serialization fast and at the same time to keep the size of the
serialized representation small.

New heartbeats contain two types of data lightweight and full version of
request/responses. The ligthweight version is reduced to simply group id
which is kept in deltafor encoded format. For now the full heartbeats
are stored in `chunked_fifo` collections and simply seriazlized with
`serde`.

Signed-off-by: Michal Maslanka <michal@redpanda.com>