cloud_storage_clients: add self configuration step
This commit introduces a new step in the start-up process of the client
pool: self configuration of the client. The intent is to give the cloud
storage client a chance to change it's configuration as it sees fits.
For instance, the ABS client will detect if the storage account has HNS
enabled and adjust the way it functions as a result.

Previously, upon start-up the client pool would wait for credentials to
be applied from upstream via the `_credentials_var` condition variable.
If any requests to create clients came before that, they would be held
until credentials had been applied. After this commit, the sequence
changes to:
1. wait for the credentials to be applied
2. perform the client self configuration step - this blocks leasing of
   clients
3. lease clients

If self configuration is not required, step 2 is a no-op.