storage: update tree to use a pointer to storage::log
Switches the tree to operate on ss::shared_ptr<log>. This is a purely
mechnical preparation for the next commit which will remove the pimpl
wrapper around. Making this change first means that the two
modifications do not need to be in the same commit. The oddities that
arise (e.g. optional<shared_ptr>) will be cleaned-up in a subsequent
commit.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit 5c9e8458d2e586077fc2990639bf28b43f1e6343)

Conflicts:
	src/v/raft/recovery_stm.cc
	src/v/storage/log_manager.cc
	src/v/storage/tests/storage_e2e_test.cc