finjector: mechanism for artificial CPU stress
This adds a mechanism for adding artificial stress to Redpanda. A
stress_payload is defined as one or multiple fibers that run in the
background, repeatedly counting up to some bounded random values/delays,
until stopped. This will be exposed via admin endpoint in a subsequent
patch.

It's expected that only one stress payload run at a time. The
public interface (stress_fiber_manager) reflects that.

The public interface exposes configuration by delay or by count. In
practice, I found the former helpful in reliably injecting reactor
stalls, and the latter helpful in having finer grained control over how
many cycles are "wasted".

This will be useful in exercising timings and reorderings of tasks that
happen when under load.