t/delta_for_test: added characterization test for deltafor
the test (and data generator) are meant to ensure that tweaks to
delta_for do not change the encoding. this is done by testing
encoding/deconding against a known good dataset, contained in
delta_for_characterization_data.

test_deltafor_characterization performs the test, while
test_deltafor_generate_characterization_data is a dummy test that
generates and prints out random characterization data (is kept here for
convenience and is marked with boost::unit_test::disabled()
to exclude it from normal runs)