r/admin_server: fixed filling up the maintenance mode information
Previously when node health report was not available its maintenance
information wasn't set at all. This way a node wasn't reported as
draining when it was unavailable. Fixed this by always relaying on the
information available in `cluster::members_table` instead of waiting for
health report. The members table is a source of truth about the
nodes maintenance mode.

Signed-off-by: Michal Maslanka <michal@redpanda.com>