tests: do not query decommissioned nodes about status
When decommissioning many nodes at the same time a test code should not
query any of the decommissioned nodes about the status at they may
contain metadata which are not updated. When node is decommissioned and
removed from the cluster it stops receiving metadata updates so its
view of the cluster state is stale.

Signed-off-by: Michal Maslanka <michal@redpanda.com>