tests/cg-recovery: retry on empty list groups res
Issues #10363 revealead that it is possible for a broker to service a
ListGroups request in a window where the leader-to-be is _becoming_ the
leader but the configuration has not replicated yet. This results in an
empty ListGroups response.

The fix is to retry the request.

Fixes #10363

Signed-off-by: NyaliaLui <nyalia@redpanda.com>
(cherry picked from commit e930db028a2676890e247516a8eb742bfb160aec)