storage: rename needs_to_roll_log condition
I found it easier to reason about the implementation when considering
whether the current segment is appendable, rather than whether we need
to roll. In a subsequent commit, I will update the condition further to
include whether the current segment has an appender, which isn't quite
as intuitive when thinking about rolling.

This patch updates the method and condition to reflect whether the
segment is appendable.