Better merging of writes in segment appender
The segment appender for a partition owns a "chunk" (default 16 KiB)
which is an in-memory buffer where append() requests end up. Some
portion of this chunk will be written when (a) it is full, so the whole
thing needs to be written or (b) a flush() call occurs in which case
only the unflushed portion (rounded to 4k boundaries) is written.

New append() calls can occur while one or more prior flush() requests
are still outstanding. Currently these are not merged effectively
in many scenarios because once a flush() call occurs previously
appended data cannot be merged with newly appended data. This is true
even if that flush hasn't been dispatched yet! That is, the flush may be
waiting for a prior write/flush to finish (we allow only one
outstanding write per chunk) meaning it is eligible for merging but
the current accounting mechanism doesn't take advantage of this
opportunity.

In a typical acks=-1 flow, append and flush requests to the appender
will be interleaved like AFAFAFAF meaning that *no* merging takes place
at all, even though the segment appender itself is capable of using one
disk flush to satisfy multiple flushes: but it ends up pipelining all
the writes (and implied flushes) and then waiting on the chunk
semaphore.

This change basically introduces a third state to in-flight writes: we
had "pending" and "done" and now "pending" is split into "queued" and
"dispatched" and we still have done. Queued means that we have decided
to issue the disk write, but we have not yet actually dispatched it,
generally because we are waiting on the chunk semaphore (i.e., there
is an earlier in-progress write to the same chunk). In this case we
can still merge new incoming appends onto this "tail" write.

Dispatched means the dma_write has occurred but the future from that
call hasn't resolved and done is when it has resolved. The meat of this
change is that the tail in-flight write can have additional writes
merged into it if it is in the "queued" state.

In the specific case of many small append + flushes this means we can
execute many fewer dma_writes and we never build a backlog of distinct
pending writes on the current chunk: we have at most one queued write
(and at most most dispatched) and any new writes for the same chunk
will merge into it.

(cherry picked from commit 0b7dd8c63516f40c5b013b545c352ec979333886)