storage/tests: Move kvstore_fixture into a header
A conflict is fixed in this where the sanitizer flag is configurable but
this change came in from two large PRs. Here to fix the conflict we just
make the sanitizer always be on for the test.

```
            1_MiB,
            config::mock_binding(10ms),
            conf.base_dir,
-            storage::debug_sanitize_files::yes);
+            storage::make_sanitized_file_config());
      },
```

(cherry picked from commit f1f925777d3ed94cdd43954af8b28cbb65e35634)