ssx: Implement deadlock detector
Deadlock detector uses semaphore names to detect incorrect locking
patters. The 'deadlock_detector_waiter_state' is supposed to be created
before waiting on the semaphore. The lifetime of the object should match
the lifetime of the acquired units object.

The detector assumes that all semaphres with the same name belong to the
same subsystem or layer. For instance, all segments are using the same
name for their internal semaphores. It also uses shard local storage to
distinguish individual fibers. The algorithm records locking order
between different semaphores (or to be precise, between
'deadlock_detector_waiter_state' objects). When the semaphore is
acquired again it checks what other sempahores the fiber has already
acquired and compares this to previous acquisitions. If the algorithm
can detect that this acquisition done in the opposite order compared to
any previous one the error is triggered.

The algorithm allows recoursive locking but generates a warning in this
case.