ducktape: Increase fetch debounce time in ManyPartitionTest
ManyPartitionTest is one of the cases which is negatively affected by
having the seperate fetch scheduling group (see
6d1223d4e4b0489c0fade40a727b9633d506398a for background).

To work around this and make the test not flaky increase the fetch
debounce timeout. This makes empty fetches less expensive and such
decreases CPU util which allows for leadership balancing to finish in
time.

I think this is better than disabling the extra group as it doesn't
fundamentally change how requests are dispatched.