cloud_storage_clients/abs: deal w/ HNS deletions
When HNS is enabled for a storage account `Delete Blob` requests will
act differently due to the file system like semantics which are
emulated.

1. Deleting a blob with a path like name only deletes the "file". For
   example, deleting "a/b/log.txt" will leave behind two blobs "a" and
   "a/b"
2 "Directory" blobs cannont be deleted with `Delete Blob` requests,
  instead one must use the ADLSv2 API.

This commit addressees the issue by sending an ADLSv2 `Filesystem
Delete` for all the valud prefix paths of the blob. For instance, if we
wish to delete "a/b/log.txt" we will send three requests sequentially to
delete "a/b/log.txt", "a/b/" and "a". If there are any other files in
the directory (i.e. we had a hash collision), the "a/b" deletion will
fail and we stop.

I'm not 100% content with this approach, but it's the only one that I
could find which was safe. ADLSv2 `Filesystem Delete` provides a
recursive delete API, so in theory we could issue just one request.
However, that's not safe when considering the fact that our prefix
hashing is not perfect.