cloud_storage_clients/abs: list only files w/ HNS
When HNS is enabled, "directory blobs" are included in the listin by
default. For instance, the container housed "a/b/log.txt", the listing
would be "a", "a/b" and "a/b/log.txt". In order to preserve the
semantics expected by callers, we use the "showonly=files" query param
for List Objects requests when HNS is enabled.