cloud_storage: Fix race in async_manifest_view_cursor
The code in the seek method calls async method
'get_materialized_manifest' to get the manifest based on query (offset).
Then, after a scheduling point it caches the start offset by querying
the STM manifest.

The following sequence of events is possible:
- Initially, there are no spillover manifests;
- 'seek' is calling 'get_materialized_manifest(0)';
- 'get_materialized_manifest' finds STM manifest which at this point
  starts with offset 0 and returns it.
- Concurrently, spillover happens and the STM manifest gets truncated.
  The segment with offset 0 gets moved to the spillover manifest.
- The 'seek' method wakes up and caches the start offset of the manifest
  which is not 0 at this point.
- The client sees the partition that starts from unexpected offset.

This commit fixes this by caching the start offset before the scheduling
point.