fix pod label causing decommission
Move pod labeling out setPodNodeIDAnnotation and into its own function.
This fixes a bug where if the pod annotation was set but the label
wasn't, the check to see that we didn't need to do any changes would be
false, allowing the next step to assume that the broker id in the
annotation was no longer valid, triggering a decommission.

By breaking this up into discreet functions, we can assure that the
annotation and label are each set correctly according to their own
logic.