cloud_storage_clients: abort on self config failure
This commit introduces logic to abort redpanda via the top level stop
signal if the self configuration of the cloud storage client failed.
For now, this is only done for ABS clients.

Self configuration should only fail if Redpanda is misconfigured.
Generally, it's a bad idea to continue running in such cases because
the disk will ultimately fill up since retention can only act after the
part of the log in question was uploaded.