rm_stm: return sequence_out_of_order only when a client misbehaves
When a request hits a former leader which still thinks it's the leader
there is a change that a valid request results in OOOSN since the broker
validates new request against the stale data. Fixing the problem by
double
checking that a node is indeed the leader.

The perf impact of linearizable_barrier doesn't matter because OOOSN is
already an exceptional situation leading to the session termination.

Fixes https://github.com/redpanda-data/redpanda/issues/12712