cloud_storage: Fix incorrect hint handling
The hint can only be used to speed up materialization of the record that
belong to the frame which was used to generate the hint. This commit
detects the situation when this is not the case and avoids using the
hint.

In order to do this the cloumn and the frame iterators are exposing the
first value of the current frame.