cloud_storage: Remove manifest lock
The manifest lock was needed because JSON serialization took long. The
manifest might have changed during the process and the result would be
invalid. Since we moved to binary format we no longer need to use this
lock. The only place where we're using this lock is admin API (the
endpoint returns JSON object). To avoid data races there this commit
copies the entire manifest before serializing it to JSON. It also
removes the lock.