r/state_machine_manager: introduced a new type to manage state machines
State machine manager is an entry point for registering state machines
built on top of replicated log. State machine managers uses a single
fiber to read and apply record batches to all managed state machines.

When a machine throws an exception or timeouts when applying batches to
its state subsequent applies are executed in the separate apply fiber
specific for that STM.

State machine manager also takes care of the snapshot consistency. It
wraps state machine snapshots in its own snapshot format which is a map
where each individual STM snapshot is stored separately. When snapshot
is
taken the manager guarantees that all the STMs have the same offset
applied so that the snapshot is consistent across the state machines.

Signed-off-by: Michal Maslanka <michal@redpanda.com>