cmake: update search swagger macro
This update synchronizes the macro with upstream seastar (this macro
isn't installed by default). The primary reason is that in soon to drop
updates to our open-source build the version from seastar via
add_subdirectory will be used and we need the tree to work both with
that model and the installation method used for vtools based builds.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>