Fix "no in-flight writes" assert
In the segment appender we assert in a few places that a particular
semaphore has 1 unit available, which is a proxy for "no in flight
writes on the last chunk". For example, when we grab max units from
the _concurrent_flushes (_cf) semaphore, we expect there to be no
concurrent writers, since other writers also obtain at least 1 unit
from this semaphore before dispatching a write.

In its current state this assert can fire. Issue #5433 describes one
way this can happen in detail, but in summary examining the units can
give false positives here: the assert may fail (0 units available) but
there are no concurrent writers because writers also try to obtain the
_cf semaphore after the _phw semaphore and may be safely blocked there.

This change tracks dispatched writes explicitly: incrementing a counter
when we dispatch a write and decrementing it after. The counter can
be incremented in the "right" place, immediately before actually
calling dma_write and after obtaining the _cf semaphore, avoiding
the problem above.

Additionally, the check is broader in that it applies to all prior
chunks, not just the most recent one. The existing check would naturally
want to check this broader condition as well, but there was no
convenient way with the way state was tracked by the appender.

In a sense, the _inflight_dispatched counter is redundant: it could be
reconstructed simply by walking the _inflight array and counting the
number of writes in DISPATCHED status. However, I am concerned that
this array could grow very large and so take a long time to traverse,
so I opted to maintain an explicit counter.

Fixes #5433.

(cherry picked from commit ea423ea903409d9a197b12f6824642bc349a13ce)