cloud_storage: add remote_file abstraction
This adds an interface to the cache for downloading and caching files.
This will be used by the controller snapshot restore path.

Why cache? While controller snapshots aren't expected to be reused
heavily, the cache itself has several nice properties built in, like
unique naming for temp files, and well-understood semantics with regards
to file cleanup.