cloud_storage: Add eager stream state
The stream state object is passed down to the hydration layer by the
segment chunk data source. It is used to signal back to the data source
if an eager stream is available for serving a fetch request, while a
download is in progress. The state object contains an ss::input_stream
which is populated by the hydration code path.