cloud_storage: Add download task to chunk data source
The download task is a background job which is used by the chunk data
source to keep track of a chunk file being downloaded from cloud
storage. While the background task is in progress, a transient stream is
used to serve data via the get() api. A download in progress must finish
before the data source is closed.