cloud_storage: Use eager stream to serve fetch request
The eager stream is populated by the hydration process. A populated
stream is then used by the chunk data source to serve get() calls. The
download started by hydration is tracked by a background download task
which must be finished before the data source is closed. There are a few
scenarios which deviate from this happy path:

* A chunk is requested and there is already a download in progress. In
  this case an eager stream will not be loaded. The data source is
  added to the wait queue and a stream will be loaded from disk once
  the download finishes.
* The chunk is found in cache - eager stream is not required, the stream
  will be loaded from the file on disk in cache.
* The download is delayed because of client acquisition or some other
  reason - the eager stream loading will be cancelled and the data
  source will wait for the stream on disk.