cloud_storage: Link reader to chunk data source
Although the chunk data source is expected to be at a different level of
abstraction than the remote segment batch reader and ideally neither
should be aware of the other, the way that the transient download
stream and the stream which writes to disk (introduced recently for
streaming reads) are coupled together means that we occasionally need
the data source to signal back to the reader that it should close the
stream it is reading from.

The transient stream and the stream which writes to disk are created
from the fanout utility, which requires that the streams are read from
consistently. If one stream is blocked (IE not consumed from), the
other can no longer read from source. So if a remote segment batch
reader is over-budget and thus stops reading from the transient stream,
the download stream is blocked, which is not acceptable.

To avoid this, the chunk data source now gets a handle to the reader. If
the data source switches to a transient stream, it signals the reader. If
the reader goes over-budget and sees that it is attached to a transient
stream, it resets its state, closing the parser and the attached stream.

The reader is re-used but this is expensive, the next time it reads, it
has to load the handle from a chunk file on disk creating a new stream
and parser.