tests: do not query decommissioned node about draining status
During node pool migration test multiple nodes are decommissioned
simultaneously. This lead to the situation in which some of those
decommissioned and already removed nodes may no longer receive metadata
updates. This caused the test to fail as it was querying those nodes for
decommission status. Fixed the problem by not querying decommissioned
node for decommission status to prevent querying stale cluster
metadata.

Fixes: #12827

Signed-off-by: Michal Maslanka <michal@redpanda.com>