tests/segment_meta_cstore_test: reproducer for fix in previous commit
the bug was caused by the false precondition that old cstore frames
(all the frame except the last one) have exactly cstore_max_frame_size
elements. When this is the case, the first element of a frame always
gets a hint associated with it, and lower_bound will retrieve a base
offset and a hint from the same frame.
but with prefix_truncate, the first frame can be recreated smaller than
cstore_max_frame_size, misaligning hints generation and the start of new
frames.
the unit test in this commit create this condition of misalignment by
creating 2 frames, truncating the first, creating a new frame and
retrieving the first element of the last frame. without the fix the
cstore.lower_bound will throw. this is easily checkable by disabling the
fix (just skip the if) in column_store::materialize()`hint_initial < hint_threshold ->
nullopt`

(cherry picked from commit 42e3736b51509c286bb7da9180af93a010b48700)