rpk: load docker host from context in rpk container
This is a workaround to new docker setups (e.g
docker desktop in Macs) which don't setup the
DOCKER_HOST env variable and does not have the
docker socket in the default location.

The alternative here is to manually parse the
docker endpoint out of the config file and the
docker context. Our current moby client does not
support contexts since this is a concept that is
only used by the CLI and the Desktop apps.