ducktape: Allow error logs in tests
The error "Can't add segment" is emitted when a duplicate segment add
command is detected.

The error "segment_reader is stuck" is now recoverable in cases of
compacted segments when the continous batch parser is at the end of
stream. When the parser is at end of stream, this log is downgraded to
info. The remaining states of the parser still emit an error.