cloud_storage: Fix timequery
If the timequery is used the async_manifest_view can chose the spillover
manifest or STM manifest incorrectly. If the timestamp is inside the gap
between the segments the async_manifest_view is not able to chose STM
manifest.

This commit fixes this by changing the timequery logic inside the
async_manifest_view. The 'in_stm' function for timequeries are
simplified and no longer needs to call partition_manifest.timequery.