c/topics_table: introduced separate exception for concurrent modifiation
When iterating over topics table we check if it was modified. The check
throws an exception made an exception specific to the situation in which
`cluster::topics_table` was modified while iterating. This way caller
can take a specific action when this situation happens.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 1783071f0615f678afc5784081dbc2193ebcc5f6)