cloud_storage: enforce read path limits
There's a number of cluster configs that impose limits on the cloud
storage read path:
* cloud_storage_max_segment_readers
* cloud_storage_max_partition_readers
* cloud_storage_max_materialized_segments_per_shard

Previously, only the partition reader limit was strictly respected.
The number of materialized segments and segment readers was allowed to
overshoot. This is detrimental in the event of a storm of cloud storage
requests since we can't trim fast enough and end up with a significant
amount of memory used by the cloud storage subsystem.

Enforcing these limits, also acts as a rudimentary multi-level
backpressure mechanism for cloud storage reads.

Note that we acquire the units pessimistically before we know that
they'll be required. This greatly simplifies the patch and avoids the
introduction of scheduling points while interacting with the partition
manifest.

(cherry picked from commit ea6b4777154fc4f7b45ee82786139bae28f75d36)