cluster: Add a plugin table
This is the per core metadata that we'll keep in memory everywhere in
the cluster.

The table supports mutations that will be used by a "plugin_backend"
that will dispatch controller commands, and a "plugin_frontend" that
will proxy the read requests from the rest of the system.

We support notifications of when data changes, so the transform
subsystem can react and update running transform processors immediately.

Also add some unit tests for this.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>