cluster: introduce plugin_frontend
plugin_frontend is the interface to transform metadata for the rest of
the system. It interfaces writing to the controller log, validating the
state changes and proxies reads to the underlying data in plugin_table.

We proxy reads to the table so that we:
- Don't expose the mutation methods in the table.
- Have a single interface to be used by the rest of the system.

This also adds an internal RPC for performing the two plugin related
commands (put and delete) on a different node if the leader is
elsewhere.

This also requires extending some error codes for the validation, which
unit tests have been written for.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>