cluster: introduce plugin_backend
This component will apply controller snapshots and updates into the
plugin tables on all cores.

Also extend the controller snapshot to have plugin related functionality
in it so the plugin_backend can read/write to it.

NOTE: This requires bumping the version number in the controller
snapshot, as we're adding a new field. As far as I know this is all
backwards compatible based on my understanding of serde. We discard
data at the end, so adding this data to the end should be fine. We only
read plugins if the old version still exists.

For clusters that are upgrading with mixed versions, the plugin snapshot
data will be ignored and we don't need to guard with a feature flag.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>