cluster: add transform validation to topics frontend
We need to ensure that when modifying topics we don't break the
invariants we need to enforce for data transforms. Namely, you have to
delete a transform before deleting attached topics, and we copartition
input and output topics. Copartitioning requires that output topics have
at least as many partitions as input topics. This invariant allows users
to increase the number of partitions in a source topic by first
increasing the number of partitions in all sink topics first.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>