handlers/produce: warning and metric for skewed timestamps
factored out validate_batch_timestamps, this function returns an
optional new timestamp to set as max_timestamp for a batch. It also
validates that the timestamps are within a window to the broker's time.
It uses the configurations introduced in the previous commit to log a
warning and increment produce_bad_timestamps, if a batch extends too far
in the past or in the future.