ducktape: Wait for spillover to stabilize before collection
When a manifest spillover happens it is uploaded to the cloud storage.
The original manifest, now truncated, is uploaded after a short period.
This means that there is a window during which the spilled over manifest
has been uploaded and the changed root manifest is not uploaded, and the
same offset range could be present in both the root and spillover in
cloud storage.

The test here constructs a reset manifest by analyzing the spillover and
root manifests. If it happens to collect these during such a window, it
ends up creating a reset manifest which is invalid, containing the same
offset range in both the segment set and the spillover set. From this
point onwards, no spillover command can succeed, because it always fails
at the overlapping range.

The change here waits for a set of manifests to appear in cloud storage
which do not have a common set of keys before creating the reset
manifest. This is done by comparing the root and spillover manifests
until a suitable combination is found.