k/group_manager: fixed handling linearizable barrier error
Previously when linearizable barrier returned an error it might lead to
situation in which a barrier was simply skipped. This in the end might
lead to a situation in which recovered consumer group state was
incomplete.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 6c6f9524d8f8acf9f066682f0563fc567baf98a9)