archival: Remove tests that check gap reupload
The tests are checking compacted reupload that starts in a gap. The
situation is impossible in redpanda since we can't have gaps in the
manifest. The tests are failing when strict checks on segment metadata
are enabled. Only alinged segment reuploads are allowed by these checks
so the reupload that starts in the gap can't really be registred by
the archival STM.

(cherry picked from commit aa2348fb3f3d790c0b41b5e25687ced269281b21)