Introduce transform::processor
This processor is the driver for a single instance of a wasm runtime.

Essentially at it's core the processor is a fiber that
read->transform->writes a partition.

There are optimizations to do to improve pipelining of the processor,
but I figured it would be better to start with the simplest thing
possible and have future changes for the optimizations.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>