r/consensus: clear consensus metrics when instance is stopped
By default metircs group uses RAII to unregister metrics. In case of the
consensus object the metrics should be cleared right after the instance
is stopped. The lifecycle of an object may be prolongated in Kafka
protocol handlers. By clearing metrics we prevent old stopped but still
present object from interfering with the newly created one.

Signed-off-by: Michal Maslanka <michal@redpanda.com>