cloud_storage: skip ahead in timequery via index
This commit updates the timequery read path to skip up to the first
index entry with a timestamp smaller than the searched one. The result
is that timequeries will hydrate/materialize a maximum of two chunks
(two because the chunk boundaries don't always line up with the entries
in the index).

If the index is not present, or was originally created in v1, the search
starts from the first chunk as it did previously.

(cherry picked from commit da8533ba5bd255fe41286e0d44013d60d93082d8)