ducktape: avoid writing empty release cache file
We noticed one of the CI runs fetched 0 releases and wrote an empty
cache file. We suspect this was because of a transient github issue.
Subsequent upgrade tests failed because they are not able to find
desired previous release in cache.

Rather than waiting for next round of tests to fail, this commit adds a
simple check that avoids writing an empty file essentially failing fast.
Next test should repopulate the file again by fetching the release
metadata from github.