sr: set authz in mitigate error
Previously the SR was setting authz in a loop at startup. This commit
sets authz during error handling.

The first place where an authz failure would occur is when the registry
creates the `_schemas` topic at startup. So this commit also modifies
`client::create_topic()` to return an authz error.

Signed-off-by: NyaliaLui <nyalia@redpanda.com>