c/config_manager: prevent config_manager from loosing members notifications
There is a dependency between the `cluster::config_manager` and members
table. Cluster configuration status updates are not explicitly removed
from status but but `config_manager` rely on notifications from members
table. The relationship is correct since all the controller log commands
are applied in order i.e. if member was deleted after its configuration
status was updated the notification is guaranteed to be generated after
the status update hence `config_manager` can use notification to cleanup
not longer relevant statuses.

Issue leading to discrepancy between members table content and
`config_manager` status map was caused by the fact that manage
notification was register after the update was generated.

Move the notification subscription earlier (to the constructor of
`config_manager`. The solution isn't perfect as there is a dependency in
order of starup and notification registration. In future we are going
to replace the notification between the two components by simply
listening to members related commands in the `config_manager` (currently
it is impossible due to limitation in `mux_state_machine`).

Signed-off-by: Michal Maslanka <michal@redpanda.com>